+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.1] - 2024-05-03
 ### Fixed
 - Correct the interfaces iterator logic to send the correct device
   introspection [#334](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/334)